Add files via upload
Adding a program intended as an exercise in understanding dictionary KEY-VALUE manipulation. The SCRUB data feature is ugly, I'll admit, but it seems to work. 

Note that I've included some sample KEY-VALUE pairs in the dictionary so one can start experimenting with the script without taking the time to build one from scratch. Most practical implementations of dictionaries would have KEY-VALUES parsed from a database or data scraped from the Internet, but for learning purposes, a dictionary included as part of the code works here.